home work

+++ 2025-09-18/snmp_v2_4_get_all.py
@@ -27,59 +27,55 @@ import asyncio
 
 # SNMP v2 获取信息汇总
 def snmpv2_get_all(ip_address, community):
+    
+   
     # 主机名
     sys_name = asyncio.run(snmpv2_get(ip_address, community, "1.3.6.1.2.1.1.5.0", port=161))[1]
 
-    # cpmCPUTotal5min
+    # cpmCPUTotal5min, int
     # 实体5分钟平均的CPU使用率 取值范围：0～100 缺省值：0
-    cpu_usage = int(asyncio.run(snmpv2_get(ip_address,
-                                           community,
-                                           "1.3.6.1.4.1.9.9.109.1.1.1.1.8.7",
-                                           port=161))[1])
+    cpu_usage = int(asyncio.run(snmpv2_get(ip_address, community, "1.3.6.1.4.1.9.9.109.1.1.1.1.8.7",port=161))[1])
+    
     # print(cpu_usage)
-    # cpmCPUMemoryUsed
-    mem_used = int(asyncio.run(snmpv2_get(ip_address,
-                                          community,
-                                          "1.3.6.1.4.1.9.9.109.1.1.1.1.12.7",
-                                          port=161))[1])
-    # cpmCPUMemoryFree
-    mem_free = int(asyncio.run(snmpv2_get(ip_address,
-                                          community,
-                                          "1.3.6.1.4.1.9.9.109.1.1.1.1.13.7",
-                                          port=161))[1])
+    # cpmCPUMemoryUsed, int
+    mem_used = int(asyncio.run(snmpv2_get(ip_address,community,"1.3.6.1.4.1.9.9.109.1.1.1.1.12.7",  port=161))[1])
+    # cpmCPUMemoryFree, int
+    mem_free = int(asyncio.run(snmpv2_get(ip_address, community, "1.3.6.1.4.1.9.9.109.1.1.1.1.13.7", port=161))[1])
 
-    mem_percent = round((mem_used / (mem_used + mem_free)*100), 4)
+    # memory utilization in percenage, float
+    mem_percent = float(round((mem_used / (mem_used + mem_free)*100), 4))
 
-    # print(mem_percent)
 
     # -----------------------------------------------------
-    # 接口名称
+    # 接口名称, list of str
     raw_name_list = asyncio.run(snmpv2_getbulk(ip_address, community, "1.3.6.1.2.1.2.2.1.2", port=161))
     if_name_list = [raw_if_name[1] for raw_if_name in raw_name_list]
-    # print(if_name_list)
+    #print(if_name_list)
 
-    # 接口速率
+    # 接口速率, list of int
     raw_speed_list = asyncio.run(snmpv2_getbulk(ip_address, community, "1.3.6.1.2.1.2.2.1.5", port=161))
     if_speed_list = [raw_speed[1] for raw_speed in raw_speed_list]
-    # print(if_speed_list)
+    #print(if_speed_list)
 
-    # 获取接口管理状态
+    # 获取接口管理状态, list of int, 1 is up, 0 is down
     raw_status_list = asyncio.run(snmpv2_getbulk(ip_address, community, "1.3.6.1.2.1.2.2.1.7", port=161))
     if_status_list = [raw_status[1] for raw_status in raw_status_list]
-    # print(if_status_list)
+    #print(if_status_list)
 
-    # 进接口字节数
+    # 入接口字节数, list of int
     raw_in_bytes_list = asyncio.run(snmpv2_getbulk(ip_address, community, "1.3.6.1.2.1.2.2.1.10", port=161))
     if_in_bytes_list = [raw_in_bytes[1] for raw_in_bytes in raw_in_bytes_list]
-    # print(if_in_bytes_list)
+    #print(if_in_bytes_list)
 
-    # 出接口字节数
+    # 出接口字节数, list of int
     raw_out_bytes_list = asyncio.run(snmpv2_getbulk(ip_address, community, "1.3.6.1.2.1.2.2.1.16", port=161))
     if_out_bytes_list = [raw_out_bytes[1] for raw_out_bytes in raw_out_bytes_list]
-    # print(if_out_bytes_list)
+    #print(if_out_bytes_list)
+
 
     interface_list = []
-    for name, speed, status, in_bytes, out_bytes in zip(if_name_list, if_speed_list, if_status_list, if_in_bytes_list, if_out_bytes_list):
+    for name, speed, status, in_bytes, out_bytes in zip(if_name_list, if_speed_list, if_status_list, if_in_bytes_list, if_out_bytes_list): # zip as concated list
+        #print(name,speed, status, in_bytes, out_bytes)
         if int(in_bytes) and int(out_bytes):
             interface_list.append({
                 'interface_name': name,
